--- packages/board-core/src/constants.ts
@@ -1,26 +1,0 @@
-import { CanvasStyles } from './types';
-
-export const defaultWindowStyles = {
-    width: 1024,
-    height: 640,
-} as const;
-
-export const defaultCanvasStyles: CanvasStyles = {
-    width: 'fit-content',
-    height: 'fit-content',
-    marginLeft: 'auto',
-    marginRight: 'auto',
-    marginBottom: 'auto',
-    marginTop: 'auto',
-    paddingLeft: '0px',
-    paddingRight: '0px',
-    paddingBottom: '0px',
-    paddingTop: '0px',
-} as const;
-
-export const defaultEnvironmentProperties = {
-    windowWidth: defaultWindowStyles.width,
-    windowHeight: defaultWindowStyles.height,
-    canvasMargin: {},
-    canvasPadding: {},
-};
