+++ packages/board-core/src/create-renderable-base.ts
@@ -1,5 +1,12 @@
 import type { IRenderableMetadataBase } from './types';
+import { createMetadata, OmitGeneralMetadata } from './create-metadata';
+import { setupBoardStage } from './setup-stage';
 import { callHooks } from './hooks';
+
+export type OmitIRenderableMetadataBase<DATA extends IRenderableMetadataBase> = Omit<
+    OmitGeneralMetadata<DATA>,
+    'setupStage' | 'cleanupStage'
+>;
 
 export async function baseRender<DATA extends IRenderableMetadataBase>(
     data: DATA,
@@ -10,4 +17,16 @@ export async function baseRender<DATA extends IRenderableMetadataBase>(
     const cleanup = await render(canvas);
     callHooks<IRenderableMetadataBase, 'afterRender'>(data, 'afterRender', canvas);
     return cleanup;
+}
+
+export function createRenderableBase<DATA extends IRenderableMetadataBase>(
+    data: OmitIRenderableMetadataBase<DATA>,
+): DATA {
+    const res: DATA = createMetadata({
+        ...data,
+        setupStage(parentElement = document.body) {
+            return setupBoardStage(this, parentElement);
+        },
+    } as DATA);
+    return res;
 }
