+++ packages/board-core/src/index.ts
@@ -2,4 +2,3 @@ export * from './create-metadata';
 export * from './create-renderable-base';
 export * from './hooks';
 export * from './types';
-export * from './constants';
