+++ packages/board-core/src/setup-stage.tsx
@@ -1,12 +1,30 @@
-import {
-    BoardSetupStageFunction,
-    ICanvasEnvironmentProps,
-    IWindowEnvironmentProps,
-    callHooks,
-    defaultCanvasStyles,
-    defaultWindowStyles,
-} from '@wixc3/board-core';
-import { IReactBoard } from './types';
+import { callHooks } from './hooks';
+import type { BoardSetupStageFunction, IWindowEnvironmentProps, ICanvasEnvironmentProps, CanvasStyles } from './types';
+
+export const defaultWindowStyles = {
+    width: 1024,
+    height: 640,
+} as const;
+
+export const defaultCanvasStyles: CanvasStyles = {
+    width: 'fit-content',
+    height: 'fit-content',
+    marginLeft: 'auto',
+    marginRight: 'auto',
+    marginBottom: 'auto',
+    marginTop: 'auto',
+    paddingLeft: '0px',
+    paddingRight: '0px',
+    paddingBottom: '0px',
+    paddingTop: '0px',
+} as const;
+
+export const defaultEnvironmentProperties = {
+    windowWidth: defaultWindowStyles.width,
+    windowHeight: defaultWindowStyles.height,
+    canvasMargin: {},
+    canvasPadding: {},
+};
 
 const applyStylesToWindow = (windowStyles: IWindowEnvironmentProps = {}, previousProps: IWindowEnvironmentProps) => {
     // we revert the changes to previous values when running cleanup
@@ -79,7 +97,7 @@ const applyStylesToCanvas = (canvas: HTMLDivElement, environmentProps: ICanvasEn
     Object.assign(canvas.style, canvasStyle);
 };
 
-export const setupBoardStage: BoardSetupStageFunction<IReactBoard> = (board, parentElement) => {
+export const setupBoardStage: BoardSetupStageFunction = (board, parentElement) => {
     const previousWindowEnvironmentProps: IWindowEnvironmentProps = {};
     const canvas = document.createElement('div');
     canvas.setAttribute('id', 'board-canvas');
