+++ packages/react-board/src/index.ts
@@ -3,4 +3,3 @@ export * from './define-app';
 export * from './define-app-types';
 export * from './types';
 export * from './content-slot';
-export * from './setup-stage';
