+++ packages/react-board/test/setup-stage.spec.tsx
@@ -1,7 +1,7 @@
 import { expect } from 'chai';
 import { createDisposables } from '@wixc3/create-disposables';
 import { createBoard } from '@wixc3/react-board';
-import { setupBoardStage } from '../src/setup-stage';
+import { setupBoardStage } from '@wixc3/board-core/src/setup-stage';
 
 describe('setupBoardStage', () => {
     const disposables = createDisposables();
